Revert "Pull in basedOn automatically"
This reverts commit 31c02b696f1aaf9e38e5a0e2139727770dc6a7aa.

+++ .malanya/index.js
@@ -25,11 +25,14 @@ module.exports = function (packinfo) {
 
     packname = packinfo.name;
 
-    const basedOn = JSON.parse(fs.readFileSync("../buildinfo.json")).original_datapack;
+    if (!fs.existsSync(packinfo.basedOn)) {
+        log(`Based on what? Couldn't find path: ${packinfo.basedOn}.`);
+        process.exit(1);
+    }
 
     log(`Preparing...`);
 
-    const baseFiles = pathEx.walkDir(basedOn + "/data");
+    const baseFiles = pathEx.walkDir(packinfo.basedOn + "/data");
     const outPath = path.join("out", packinfo.name);
     pathEx.removeDir("out");
     fs.mkdirSync(outPath, { recursive: true });
@@ -42,7 +45,7 @@ module.exports = function (packinfo) {
     log(`Prepared for processing in ${stopwatchLap()}ms`);
 
     for (const file of baseFiles) {
-        const newFileName = file.substring(basedOn.length)
+        const newFileName = file.substring(packinfo.basedOn.length)
         try {
             if (path.extname(file) != ".json")
                 continue;
